update kamon config
Kamon.start() needs to be called before the first actor system is allocated, otherwise an exception is thrown. Scala doesn't have static initializer blocks and objects are instantiated lazily, so the only solution i found, was to create a KamonIntializer object, and "ping" it before actor system instantiation. Calling Kamon.start() twice causes an exception, so putting Kamon.start() in every query is not an option.